+++ cliquet/tests/test_authorization.py
@@ -71,6 +71,7 @@ class RouteFactoryTest(unittest.TestCase):
     def test_attributes_are_none_with_blank_requests(self):
         request = Request.blank(path='/')
         request.registry = mock.Mock(settings={})
+        request.authn_type = 'fxa'
         context = RouteFactory(request)
         self.assertIsNone(context.object_id)
         self.assertIsNone(context.required_permission)
@@ -85,6 +86,7 @@ class RouteFactoryTest(unittest.TestCase):
         request.registry.settings = {}
 
         context = RouteFactory(request)
+        self.assertIsNone(context.prefixed_userid)
         self.assertIsNone(context.object_id)
         self.assertIsNone(context.required_permission)
         self.assertIsNone(context.resource_name)
@@ -115,6 +117,7 @@ class AuthorizationPolicyTest(unittest.TestCase):
         self.authz = AuthorizationPolicy()
         self.authz.get_bound_permissions = mock.sentinel.get_bound_perms
         self.context = mock.MagicMock()
+        self.context.prefixed_userid = None
         self.context.allowed_principals = []
         self.context.object_id = mock.sentinel.object_id
         self.context.required_permission = 'read'
@@ -164,3 +167,19 @@ class AuthorizationPolicyTest(unittest.TestCase):
             self.context, ['fxa:user'], 'dynamic')
         self.context.check_permission.assert_not_called()
         self.assertTrue(has_permission)
+
+    def test_prefixed_userid_is_added_to_principals(self):
+        self.context.prefixed_userid = 'fxa:userid'
+        self.authz.permits(self.context, self.principals, 'foobar')
+        self.context.check_permission.assert_called_with(
+            'foobar',
+            self.principals + ['fxa:userid', 'fxa_userid'],
+            get_bound_permissions=mock.sentinel.get_bound_perms)
+
+    def test_unprefixed_userid_is_removed_from_principals(self):
+        self.context.prefixed_userid = 'fxa:userid'
+        self.authz.permits(self.context, ['userid'], 'foobar')
+        self.context.check_permission.assert_called_with(
+            'foobar',
+            ['fxa:userid', 'fxa_userid'],
+            get_bound_permissions=mock.sentinel.get_bound_perms)